Move project creation and detection into reccmp module
