+++ reccmp/project/common.py
@@ -1,0 +1,3 @@
+RECCMP_PROJECT_CONFIG = "reccmp-project.yml"
+RECCMP_USER_CONFIG = "reccmp-user.yml"
+RECCMP_BUILD_CONFIG = "reccmp-build.yml"
