+++ reccmp/project/error.py
@@ -1,0 +1,18 @@
+class RecCmpProjectException(Exception):
+    pass
+
+
+class RecCmpProjectNotFoundException(RecCmpProjectException):
+    pass
+
+
+class InvalidRecCmpProjectException(RecCmpProjectException):
+    pass
+
+
+class InvalidRecCmpArgumentException(RecCmpProjectException):
+    pass
+
+
+class UnknownRecCmpTargetException(RecCmpProjectException):
+    pass
